Update priority scaling to handle values above the signed byte range. Change methods to use int so facades see original value, contain any casting to within facade.

+++ qpid-jms-client/src/main/java/org/apache/qpid/jms/message/JmsMessage.java
@@ -235,7 +235,7 @@ public class JmsMessage implements javax.jms.Message {
 
     @Override
     public void setJMSPriority(int priority) throws JMSException {
-        facade.setPriority((byte) priority);
+        facade.setPriority(priority);
     }
 
     @Override
